Create checkappendsave.ps1
Check if the file exists, check if the $name has been written. Someone wanted a quick way to check IP addresses of clients machine using a particular resource.  Change the file path to a UNC path and destribute. Not how I would do it but that's what the network architect asked us for.